requirements.txt
Django==4.1.2
#psycopg2
pytz
requests
mysqlclient
django-mysql
djangorestframework
drf-yasg
yq
drf-extensions
drf_spectacular
djangorestframework-simplejwt